Update index of client data to return partial data
The `/api/user/:user_uuid/clients/data` is intended for the frontend All Clients page.
It was origianlly returning an array of all related clients data, but the frontend will only need a few pieces of the overall client information. As we intend to add additional information (such as the 30+ workflow tasks that will be associated with a client), this will quickly prove to be a large and unnecessary network call.

To prevent such a large network data transfer, this change tweaks the index to return partial client data. To be even more specific, it will only return the keys that the frontend All Clients page will need.

New Partial Client Shape
```
[{
         "client_first_name": "Natasha",
         "current_stage": {
             "category": "New Client Inquiry",
             "is_completed": false,
             "step": "Request More Information",
             "uuid": "fe71fd1a-32a5-497c-b480-a510bf94bace",
          },
         "package_name": "Wedding Premier",
        "partner_first_name": "Zihao",
        "upcoming_shoot_date": "April 17, 2020",
        "uuid": "ce20b995-0368-4a59-9ae4-ad858b77f8af",
}]
```